+++ quaternions/__init__.py
@@ -1,3 +1,3 @@
-from quaternions.quaternions import Quaternion  # NOQA
+from quaternions.quaternion import Quaternion  # NOQA
 from quaternions.general_quaternion import GeneralQuaternion, QuaternionError  # NOQA
 from quaternions.version import __version__  # NOQA
