+++ quaternions/quaternion.py
